requirements.txt
asgiref==3.8.1
Django==5.1.3
django-environ==0.11.2
sqlparse==0.5.2
tzdata==2024.2
psycopg2-binary==2.9.10
djangorestframework==3.15.2
dj-rest-auth==7.0.0
django-allauth==65.2.0
django-allauth[socialaccount]
PyJWT==2.10.0
requests==2.32.3
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
drf-spectacular==0.27.2
pillow==11.0.0






